requirements.txt
pixivpy
opencv-python
numpy
seaborn
